# static/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Whiteflag API",
        "version": "1.0.0-alpha",
...
        "x-logo": {
            "url": "wflogo.png",
            "altText": "Whiteflag"
        }
    },
...
    "servers": [
        {
            "url": "http://{hostname}:{port}/",
            "variables": {
                "hostname": {
                    "default": "localhost",
                    "description": "Configurable hostname"
                },
                "port": {
                    "default": "5746",
                    "description": "Configurable port number"
                }
            }
        }
    ],
    "security": [
        {
            "description": "Basic HTTP authorization",
            "type": "http",
            "scheme": "basic"
        }
    ],
    "paths": {
...
        "/blockchains/{blockchain}/accounts": {
            "get": {
                "tags": [ "Blockchains", "Accounts", "State" ],
                "description": "Returns the accounts of the specified blockchain.",
                "operationId": "getAccounts",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/blockchainAccounts"
                    },
                    "404": {
                        "$ref": "#/components/responses/error404"
                    },
                    "500": {
                        "$ref": "#/components/responses/error500"
                    }
                }
            },
            "post": {
                "tags": [ "Blockchains", "Accounts" ],
                "description": "Creates a new blockchain account. This operation may be disabled in the configuration.",
                "operationId": "createAccount",
                "requestBody": {
                    "$ref": "#/components/requestBodies/createAccount"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/blockchainAccountCreated"
                    },
                    "400": {
                        "$ref": "#/components/responses/error400"
                    },
                    "500": {
                        "$ref": "#/components/responses/error500"
                    },
                    "501": {
                        "$ref": "#/components/responses/error501"
                    },
                    "503": {
                        "$ref": "#/components/responses/error503"
                    }
                }
            }
        },
        "/blockchains/{blockchain}/accounts/{account}": {
            "get": {
                "tags": [ "Blockchains", "Accounts", "State" ],
                "description": "Returns the details of the specified blockchain account.",
                "operationId": "getAccount",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/blockchainAccount"
                    },
                    "404": {
                        "$ref": "#/components/responses/error404"
                    },
                    "500": {
                        "$ref": "#/components/responses/error500"
                    }
                }
            },
            "patch": {
                "tags": [ "Blockchains", "Accounts" ],
                "description": "Updates or adds custom properties of the specified blockchain account. Please be carefull as all attributes can be changed resulting in an unusable account. This operation may be disabled in the configuration.",
                "operationId": "updateAccount",
                "requestBody": {
                    "$ref": "#/components/requestBodies/updateAccount"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/blockchainAccount"
                    },
                    "400": {
                        "$ref": "#/components/responses/error400"
                    },
                    "500": {
                        "$ref": "#/components/responses/error500"
                    },
                    "501": {
                        "$ref": "#/components/responses/error501"
                    },
                    "503": {
                        "$ref": "#/components/responses/error503"
                    }
                }
            },
            "delete": {
                "tags": [ "Blockchains", "Accounts" ],
                "description": "Deletes the specified blockchain account. Be carefull as all data will be unrecoverably lost. This operation may be disabled in the configuration.",
                "operationId": "deleteAccount",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/blockchainAccount"
                    },
                    "400": {
                        "$ref": "#/components/responses/error400"
                    },
                    "500": {
                        "$ref": "#/components/responses/error500"
                    },
                    "501": {
                        "$ref": "#/components/responses/error501"
                    },
                    "503": {
                        "$ref": "#/components/responses/error503"
                    }
                }
            }
        },
        "/blockchains/{blockchain}/accounts/{account}/sign": {
            "post": {
                "tags": [ "Blockchains", "Accounts", "Signatures" ],
                "description": "Signs a Whiteflag authentication payload for the blockchain address specified in the payload and returns the resulting signature. This operation may be disabled in the configuration.",
                "operationId": "createSignature",
                "requestBody": {
                    "$ref": "#/components/requestBodies/wfSignaturePayload"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/wfSignature"
                    },
                    "400": {
                        "$ref": "#/components/responses/error400"
                    },
                    "404": {
                        "$ref": "#/components/responses/error404"
                    },
                    "500": {
                        "$ref": "#/components/responses/error500"
                    },
                    "501": {
                        "$ref": "#/components/responses/error501"
                    },
                    "503": {
                        "$ref": "#/components/responses/error503"
                    }
                }
            }
        },
        "/blockchains/{blockchain}/accounts/{account}/transfer": {
            "post": {
                "tags": [ "Blockchains", "Accounts" ],
                "description": "Transfers value to another blockchain account. This operation may be disabled in the configuration.",
                "operationId": "transferValue",
                "requestBody": {
                    "$ref": "#/components/requestBodies/transferValue"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/transferValue"
                    },
                    "400": {
                        "$ref": "#/components/responses/error400"
                    },
                    "404": {
                        "$ref": "#/components/responses/error404"
                    },
                    "500": {
                        "$ref": "#/components/responses/error500"
                    },
                    "501": {
                        "$ref": "#/components/responses/error501"
                    },
                    "503": {
                        "$ref": "#/components/responses/error503"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "schemas": {
...
            "responseBodyMetaObject": {
                "title": "API response metadata",
                "description": "Meta data about the processing of the request",
                "type": "object",
                "properties": {
                    "additionalProperties": true,
                    "operationId": {
                        "description": "Unique string defined by the OpenAPI defintion used to identify the operation",
                        "type": "string"
                    },
                    "request": {
                        "description": "Details of the request on the endpoint",
                        "type": "object",
                        "properties": {
                            "client": {
                                "description": "The ip address of the client that made the request",
                                "type": "string"
                            },
                            "method": {
                                "description": "The HTTP method used to make the request",
                                "type": "string"
                            },
                            "endpoint": {
                                "description": "The endpoint on which the request has been made",
                                "type": "string"
                            }
                        }
                    },
                    "info": {
                        "description": "Information about how the data was processed",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "warnings": {
                        "description": "Warnings generated when processing the request",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "errors": {
                        "description": "Errors generated when processing the request",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "responseBodyErrorObject": {
                "title": "API response errors",
                "description": "Errors describing why the request could not succesfully be fulfilled",
                "type": "array",
                "items": {
                    "type": "string"
                }
            },
            "responseBodyErrors": {
                "type": "object",
                "properties": {
                    "meta": {
                        "$ref": "#/components/schemas/responseBodyMetaObject"
                    },
                    "errors": {
                        "$ref": "#/components/schemas/responseBodyErrorObject"
                    }
                }
            }
        },
        "requestBodies": {
...
            "wfSignaturePayload": {
                "description": "Payload for a Whiteflag Authentication Signature",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "protocol/signature.schema.json#/definitions/wfSignaturePayload"
                        }
                    }
                }
            },
...
            "transferValue": {
                "description": "Request to transfer value to another blockchain account",
                "content": {
                    "application/json": {
                        "schema": {
                            "title": "Blockchain value transfer request body",
                            "description": "Blockchain value transfer request body",
                            "type": "object",
                            "additionalProperties": true,
                            "properties": {
                                "fromAddress": {
                                    "description": "The address of the blockchain account to transfer value from",
                                    "type": "string"
                                },
                                "toAddress": {
                                    "description": "The address of the blockchain account to transfer value to",
                                    "type": "string"
                                },
                                "value": {
                                    "description": "Value to be transferred in the main currency of the blockchain",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "createAccount": {
                "description": "Request to create blockchain account",
                "content": {
                    "application/json": {
                        "schema": {
                            "description": "Blockchain account creation data object",
                            "type": "object",
                            "properties": {
                                "privateKey": {
                                    "description": "Optional private key in raw hexadecimal format to create account with",
                                    "type": "string",
                                    "pattern": "[a-fA-F0-9]"
                                }
                            }
                        }
                    }
                }
            },
            "updateAccount": {
                "description": "Request to update blockchain account",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "protocol/state.schema.json#/definitions/blockchainAccountObject"
                        }
                    }
                }
            },
...
        },
        "responses": {
            "error400": {
                "description": "Invalid request, typically because of a malformed syntax or protocol error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/responseBodyErrors"
                        }
                    }
                }
            },
...
            "error404": {
                "description": "Requested resource not found or no data available",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/responseBodyErrors"
                        }
                    }
                }
            },
            "error500": {
                "description": "Internal error preventing the API to process the request",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/responseBodyErrors"
                        }
                    }
                }
            },
            "error501": {
                "description": "Function not implemented, such as a missing protocol feature or not implemented blockchain",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/responseBodyErrors"
                        }
                    }
                }
            },
            "error503": {
                "description": "Function currently not available, such as a blockchain connection that is not enabled",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/responseBodyErrors"
                        }
                    }
                }
            },
...
            "wfSignature": {
                "description": "Succesfully processed the Whiteflag authentication signature and returning the result",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta": {
                                    "$ref": "#/components/schemas/responseBodyMetaObject"
                                },
                                "data": {
                                    "$ref": "protocol/signature.schema.json#/definitions/wfSignature"
                                }
                            }
                        }
                    }
                }
            },
...
            "blockchainAccounts": {
                "description": "Successfully retrieved and returning all blockchain accounts",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta": {
                                    "$ref": "#/components/schemas/responseBodyMetaObject"
                                },
                                "data": {
                                    "description": "Blockchain accounts",
                                    "type": "array",
                                    "items": {
                                        "$ref": "protocol/state.schema.json#/definitions/blockchainAccountObject"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "blockchainAccountCreated": {
                "description": "Successfully created the blockchain account, and returning account data",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta": {
                                    "$ref": "#/components/schemas/responseBodyMetaObject"
                                },
                                "data": {
                                    "$ref": "protocol/state.schema.json#/definitions/blockchainAccountObject"
                                }
                            }
                        }
                    }
                }
            },
            "blockchainAccount": {
                "description": "Successfully retrieved, updated or deleted the blockchain account, and returning account data",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta": {
                                    "$ref": "#/components/schemas/responseBodyMetaObject"
                                },
                                "data": {
                                    "$ref": "protocol/state.schema.json#/definitions/blockchainAccountObject"
                                }
                            }
                        }
                    }
                }
            },
            "transferValue": {
                "description": "Successfully transferred value to another blockchain account and returning the transaction hash",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta": {
                                    "$ref": "#/components/schemas/responseBodyMetaObject"
                                },
                                "data": {
                                    "description": "Value transfer result",
                                    "type": "object",
                                    "properties": {
                                        "transactionHash": {
                                            "description": "The transaction hash of the value transfer",
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
...
        }
    }
}